Refactor archive description, remove deprecated items (meta, hardcoded properties)
